move compUpdateLifeVar and compUpdateTreeLife to separate files for legacy and non-legacy case
